Add 'user' session field in AutoSignIn and LoginPost
To know if the credentials of the loggedin user were not changed, we
used to hit the database on each request with data from cookie.

To prevent database roundtrips when the user is logged, we now
cache the user in session store. This has a side effect which enables
users to continue using their session when they have changed their
password elsewhere (on ERP for example). This might be ok for now but
will more likely change in the future.